# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CEPS Portal API — referral intake, authorizations, invoices, payments, remittances, vendors, reporting
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients (scoped by role)
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Client" }
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Client detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404": { description: Not found }
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client (staff; incl. coordinator reassignment)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Soft-delete a client (staff)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "404": { description: Not found }
  /clients/{id}/case:
    get:
      operationId: getClientCase
      tags: [clients]
      summary: Full case view — referrals, authorizations, invoices, payments, remittances
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Case bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientCase"
...
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fees (scoped by role)
      parameters:
        - name: clientId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Fees
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Fee" }
    post:
      operationId: createFee
      tags: [fees]
      summary: Manually create/adjust a fee (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fee"
  /fees/{id}:
    patch:
      operationId: updateFee
      tags: [fees]
      summary: Update a fee (staff)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeUpdateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fee"
    delete:
      operationId: deleteFee
      tags: [fees]
      summary: Soft-delete a fee (staff)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "404": { description: Not found }
...
components:
  schemas:
...
    OkResult:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    Client:
      type: object
      required: [id, firstName, lastName, dateOfBirth, uciNumber, status]
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        dateOfBirth: { type: string }
        uciNumber: { type: string }
        address: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        status: { type: string, enum: [active, inactive, closed] }
        regionalCenter: { type: ["string", "null"] }
        preferredLanguage: { type: ["string", "null"] }
        assignedCoordinatorId: { type: ["string", "null"] }
        assignedCoordinatorName: { type: ["string", "null"] }
        isMinor: { type: ["boolean", "null"] }
        familyRepName: { type: ["string", "null"] }
        familyRepPhone: { type: ["string", "null"] }
        familyRepEmail: { type: ["string", "null"] }
        familyRepAddress: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"] }
    ClientInput:
      type: object
      required: [firstName, lastName, dateOfBirth, uciNumber]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        dateOfBirth: { type: string }
        uciNumber: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        status: { type: string, enum: [active, inactive, closed] }
        regionalCenter: { type: string }
        preferredLanguage: { type: string }
        isMinor: { type: boolean }
        familyRepName: { type: string }
        familyRepPhone: { type: string }
        familyRepEmail: { type: string }
        familyRepAddress: { type: string }
    ClientUpdate:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        dateOfBirth: { type: string }
        uciNumber: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        status: { type: string, enum: [active, inactive, closed] }
        regionalCenter: { type: string }
        preferredLanguage: { type: string }
        assignedCoordinatorId: { type: ["string", "null"] }
        isMinor: { type: boolean }
        familyRepName: { type: string }
        familyRepPhone: { type: string }
        familyRepEmail: { type: string }
        familyRepAddress: { type: string }
    ClientCase:
      type: object
      required: [client, referrals, authorizations, invoices, payments, remittances]
      properties:
        client: { $ref: "#/components/schemas/Client" }
        referrals:
          type: array
          items: { $ref: "#/components/schemas/Referral" }
        authorizations:
          type: array
          items: { $ref: "#/components/schemas/Authorization" }
        invoices:
          type: array
          items: { $ref: "#/components/schemas/Invoice" }
        payments:
          type: array
          items: { $ref: "#/components/schemas/Payment" }
        remittances:
          type: array
          items: { $ref: "#/components/schemas/Remittance" }
    ReferralIntakeFields:
      type: object
      description: Structured referral form data (Section 9 field list)
      properties:
        regionalCenterName: { type: string }
        coordinatorName: { type: string }
        coordinatorEmail: { type: string }
        coordinatorPhone: { type: string }
        vendorAcceptsChecks: { type: boolean }
        vendorName: { type: string }
        vendorEmail: { type: string }
        vendorPhone: { type: string }
        vendorContactPerson: { type: string }
        vendorServiceStreet: { type: string }
        vendorServiceCity: { type: string }
        vendorServiceZip: { type: string }
        vendorServiceState: { type: string }
        vendorBillingDifferent: { type: string, enum: [yes, no, unknown] }
        vendorBillingStreet: { type: string }
        vendorBillingCity: { type: string }
        vendorBillingZip: { type: string }
        vendorBillingState: { type: string }
        serviceType: { type: string, enum: [direct_pay_459, reimbursement_024] }
        activityDescription: { type: string }
        serviceStartDate: { type: string }
        serviceEndDate: { type: string }
        posNumber: { type: string }
        posStartDate: { type: string }
        posEndDate: { type: string }
        clientFirstName: { type: string }
        clientLastName: { type: string }
        clientDob: { type: string }
        clientUci: { type: string }
        preferredLanguage: { type: string }
        familyRepName: { type: string }
        clientIsMinor: { type: boolean }
        contactPhone: { type: string }
        contactEmail: { type: string }
        contactStreet: { type: string }
        contactCity: { type: string }
        contactZip: { type: string }
        contactState: { type: string }
    Referral:
      type: object
      required: [id, clientId, referralDate, status]
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: ["string", "null"] }
        serviceCoordinatorId: { type: ["string", "null"] }
        coordinatorName: { type: ["string", "null"] }
        referralDate: { type: string }
        status:
          type: string
          enum: [intake, pending_signature, pending_auth, pending_w9, pending_invoice, active, closed]
        submittedVia: { type: ["string", "null"], enum: [portal, staff_manual_entry, null] }
        intakeFields: { $ref: "#/components/schemas/ReferralIntakeFields" }
        parentEmail: { type: ["string", "null"] }
        parentSignedAt: { type: ["string", "null"] }
        signedByName: { type: ["string", "null"] }
        altaAuthReceivedAt: { type: ["string", "null"] }
        serviceFrequency: { type: ["string", "null"], enum: [one_time, monthly, null] }
        notes: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"] }
...
    Authorization:
      type: object
      required: [id, clientId, authNumber, serviceCode, paymentType, servicePeriodStart, servicePeriodEnd, maxPeriodAmount, status]
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: ["string", "null"] }
        vendorId: { type: ["string", "null"] }
        vendorName: { type: ["string", "null"] }
        authNumber: { type: string }
        serviceCode: { type: string, enum: ["459", "024", "490"] }
        paymentType: { type: string, enum: [direct_payment, reimbursement, fee] }
        activityDescription: { type: ["string", "null"] }
        servicePeriodStart: { type: string }
        servicePeriodEnd: { type: string }
        monthlyAmount: { type: ["string", "null"] }
        oneTimeAmount: { type: ["string", "null"] }
        maxPeriodAmount: { type: string }
        units: { type: ["integer", "null"] }
        status: { type: string, enum: [active, expired, pending, exhausted] }
        posPdfUrl: { type: ["string", "null"] }
        receivedDate: { type: ["string", "null"] }
        totalPaid: { type: ["string", "null"] }
        remainingAmount: { type: ["string", "null"] }
        daysUntilExpiry: { type: ["integer", "null"] }
...
    Invoice:
      type: object
      required: [id, clientId, submittedByRole, submittedDate, serviceMonth, amountRequested, paymentType, status]
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: ["string", "null"] }
        authorizationId: { type: ["string", "null"] }
        authNumber: { type: ["string", "null"] }
        vendorId: { type: ["string", "null"] }
        vendorName: { type: ["string", "null"] }
        submittedByRole: { type: string, enum: [vendor, parent, staff] }
        submittedDate: { type: string }
        serviceMonth: { type: string, description: "YYYY-MM" }
        amountRequested: { type: string }
        paymentType: { type: string, enum: [direct_payment, reimbursement] }
        documentUrl: { type: ["string", "null"] }
        status: { type: string, enum: [pending_review, validated, approved, rejected, duplicate] }
        reviewedBy: { type: ["string", "null"] }
        reviewedByName: { type: ["string", "null"] }
        reviewedAt: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"] }
...
    Payment:
      type: object
      required: [id, clientId, qbCheckNumber, checkDate, amount, paymentType, source]
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: ["string", "null"] }
        authorizationId: { type: ["string", "null"] }
        authNumber: { type: ["string", "null"] }
        vendorId: { type: ["string", "null"] }
        vendorName: { type: ["string", "null"] }
        invoiceId: { type: ["string", "null"] }
        qbCheckNumber: { type: string }
        checkDate: { type: string }
        amount: { type: string }
        paymentMonth: { type: ["string", "null"] }
        paymentType: { type: string, enum: [direct_payment, reimbursement, fee] }
        source: { type: string, enum: [quickbooks, manual] }
        loggedBy: { type: ["string", "null"] }
        remitted: { type: ["boolean", "null"] }
        createdAt: { type: ["string", "null"] }
...
    Remittance:
      type: object
      required: [id, clientId, remittanceDate, amount, status, source, autoMatched]
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: ["string", "null"] }
        authorizationId: { type: ["string", "null"] }
        authNumber: { type: ["string", "null"] }
        altaReference: { type: ["string", "null"] }
        remittanceDate: { type: string }
        amount: { type: string }
        paymentMonth: { type: ["string", "null"] }
        status: { type: string, enum: [pending, received, matched] }
        source: { type: string, enum: [alta_regional, manual] }
        matchedPaymentId: { type: ["string", "null"] }
        autoMatched: { type: boolean }
        remittanceBatchId: { type: ["string", "null"] }
...
    Fee:
      type: object
      required: [id, clientId, amount, status]
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: ["string", "null"] }
        paymentId: { type: ["string", "null"] }
        authorizationId: { type: ["string", "null"] }
        amount: { type: string }
        ruleApplied: { type: ["string", "null"] }
        status: { type: string, enum: [pending, invoiced, collected, waived] }
        notes: { type: ["string", "null"] }
        createdBy: { type: ["string", "null"] }
        createdAt: { type: ["string", "null"] }
    FeeInput:
      type: object
      required: [clientId, amount]
      properties:
        clientId: { type: string }
        paymentId: { type: ["string", "null"] }
        authorizationId: { type: ["string", "null"] }
        amount: { type: string }
        ruleApplied: { type: string }
        status: { type: string, enum: [pending, invoiced, collected, waived] }
        notes: { type: string }
    FeeUpdateInput:
      type: object
      properties:
        amount: { type: string }
        status: { type: string, enum: [pending, invoiced, collected, waived] }
        notes: { type: string }
...